app/src/main/java/com/gym/app/sms/service/SMSService.java
package com.gym.app.sms.service;

import com.gym.app.dto.OtpRequest;
import com.gym.app.dto.OtpResponseDto;
import com.gym.app.dto.OtpValidationRequest;
import com.gym.app.enums.OtpStatus;
import com.gym.app.security.config.TwilioConfig;
import com.gym.app.sms.SMSRequest;
import com.twilio.Twilio;
import com.twilio.rest.api.v2010.account.Message;
import com.twilio.type.PhoneNumber;
import jakarta.annotation.PostConstruct;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.text.DecimalFormat;
import java.util.Map;
import java.util.Random;
import java.util.Set;
import java.util.concurrent.ConcurrentHashMap;

@Service
@Slf4j
public class SMSService {

    @Autowired
    private TwilioConfig twilioConfig;

    private final Map<String, String> otpMap = new ConcurrentHashMap<>();

    @PostConstruct
    public void initTwilio() {
        Twilio.init(twilioConfig.getAccountSid(), twilioConfig.getAuthToken());
    }

    public OtpResponseDto sendSMS(SMSRequest smsRequest) {
        OtpResponseDto otpResponseDto = null;
        try {
            PhoneNumber to = new PhoneNumber(smsRequest.getDestinationSMSNumber()); // to
            PhoneNumber from = new PhoneNumber(twilioConfig.getTrialNumber()); // from
            String otp = generateOTP();
            String otpMessage = smsRequest.getSmsMessage();
            Message message = Message.creator(to, from, otpMessage).create();
            otpMap.put(smsRequest.getSmsMessage(), otp);
            otpResponseDto = new OtpResponseDto(OtpStatus.DELIVERED, otpMessage);
        } catch (Exception e) {
            e.printStackTrace();
            otpResponseDto = new OtpResponseDto(OtpStatus.FAILED, e.getMessage());
        }
        return otpResponseDto;
    }

    public String validateOtp(OtpValidationRequest otpValidationRequest) {
        Set<String> keys = otpMap.keySet();
        String username = null;
        for (String key : keys)
            username = key;
        if (otpValidationRequest.getUsername().equals(username)) {
            otpMap.remove(username, otpValidationRequest.getOtpNumber());
            return "OTP is valid!";
        } else {
            return "OTP is invalid!";
        }
    }

    private String generateOTP() {
        return new DecimalFormat("000000").format(new Random().nextInt(999999));
    }
}